Split singleton methods into a trait
